requirements.txt
click==6.6
Django==1.8.4
django-rest-swagger==0.3.4
django-suit==0.2.15
django-tinymce==2.3.0
djangorestframework==3.2.4
docutils==0.12
Flask==0.11.1
gcm==0.3
gunicorn==19.5.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
MySQL-python==1.2.5
Pillow==2.9.0
PyYAML==3.11
reportlab==3.3.0
requests==2.9.1
sorl-thumbnail==12.3
suds==0.4
suit==2.0.2
Werkzeug==0.11.10
